+++ examples/android/app/src/main/java/io/bigmoeonedge/example/SettingsScreen.kt
@@ -14,8 +14,18 @@ import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 
 /**
- * Groups every tunable the engine exposes. Changes are applied to [current] live and
- * persisted by the caller. The layout mirrors the CLI flags one-to-one.
+ * Every tunable the engine exposes, grouped by what it is for. Changes apply to [current] live and
+ * the caller persists them.
+ *
+ * Each category shows the recommended configuration first and folds the rest into an
+ * [ExperimentalGroup]. That is a statement about evidence, not about how finished the code is:
+ * inside are the levers measured on one device, or measured once, or still owed a measurement.
+ * They stay in the release build because testing them on other hardware is what the demo app is
+ * for, and a lever nobody can reach is a lever nobody can refute.
+ *
+ * Descriptions say what a setting does for the person reading, not how it is implemented, and
+ * carry no measured figures. A number here would need the device, the model and the day beside it
+ * to mean anything, and none of those fit under a switch. The documents do that job.
  */
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -38,6 +48,9 @@ fun SettingsScreen(current: AppSettings, onChange: (AppSettings) -> Unit, onBack
             )
         },
     ) { padding ->
+        val stream = !current.mmap
+        val cacheOn = current.cacheMb == AppSettings.CACHE_AUTO || current.cacheMb > 0
+
         Column(
             Modifier
                 .padding(padding)
@@ -50,227 +63,212 @@ fun SettingsScreen(current: AppSettings, onChange: (AppSettings) -> Unit, onBack
                 // mmap is the no-streaming baseline. When on, every streaming knob below is
                 // inert (the CLI omits --moe-stream and all sub-flags), so they are disabled.
                 SwitchRow(
-                    "mmap baseline (no streaming)",
-                    "Load the model the ordinary way, without streaming experts. The baseline to compare against",
+                    "Load the whole model instead",
+                    "Let the system load the model the ordinary way rather than streaming experts " +
+                        "from storage. On a model past your memory this is much slower; it is here " +
+                        "to compare against.",
                     current.mmap,
                 ) { onChange(current.copy(mmap = it)) }
 
-                val stream = !current.mmap
-                val cacheOn = current.cacheMb == AppSettings.CACHE_AUTO || current.cacheMb > 0
                 IntSetting(
-                    "Expert cache (MiB)", AppSettings.CACHE_CHOICES, current.cacheMb,
+                    "Expert cache", AppSettings.CACHE_CHOICES, current.cacheMb,
                     format = {
                         when (it) {
                             AppSettings.CACHE_AUTO -> "Auto"
-                            0 -> "off"
+                            0 -> "Off"
                             else -> "$it MiB"
                         }
                     },
                     enabled = stream,
                 ) { onChange(current.copy(cacheMb = it)) }
-                Text(
-                    "Experts already in memory cost no read at all, so a bigger cache means less waiting on " +
-                        "flash — paid for in memory the rest of the system no longer has. Auto sizes it once " +
-                        "at load from what is free, then holds that budget. " +
-                        if (AppSettings.cacheNeedsForce(current.cacheMb))
-                            "The smallest rungs sit below the engine's own floor: a cache too small to hold " +
-                                "one token's experts evicts them before they are reused, and only churns."
-                        else "The smallest rungs sit below the engine's floor and have to be forced.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
+                Hint(
+                    "An expert already in memory costs no read at all, so a bigger cache means less " +
+                        "waiting on storage, paid for in memory the rest of the system no longer has. " +
+                        "Auto sizes it once when the model opens and then holds that budget. " +
+                        "The smallest sizes sit below the engine's own floor: a cache too small to " +
+                        "hold what a single token needs throws experts away before they are reused, " +
+                        "so it only churns."
                 )
                 IntSetting(
-                    "Auto cache ceiling (MiB)", AppSettings.CACHE_CEIL_CHOICES, current.cacheCeilMb,
-                    format = { if (it == 0) "no cap" else "$it MiB" },
+                    "Cache ceiling for Auto", AppSettings.CACHE_CEIL_CHOICES, current.cacheCeilMb,
+                    format = { if (it == 0) "No cap" else "$it MiB" },
                     enabled = stream && current.cacheMb == AppSettings.CACHE_AUTO,
                 ) { onChange(current.copy(cacheCeilMb = it)) }
-                Text(
-                    "Caps what Auto may claim. The system reports the model's own mapped weights as free " +
-                        "memory, so left uncapped Auto can ask for more than really exists.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
+                Hint(
+                    "Caps what Auto may claim. The system counts the model's own mapped weights as " +
+                        "free memory, so left uncapped Auto can ask for more than really exists."
                 )
-                IntSetting("Parallel I/O lanes", AppSettings.IO_CHOICES, current.ioThreads, enabled = stream) {
+                IntSetting("Parallel reads", AppSettings.IO_CHOICES, current.ioThreads, enabled = stream) {
                     onChange(current.copy(ioThreads = it))
                 }
-                Text(
-                    "How many expert reads are in flight at once. More lanes help only until the flash itself is saturated.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
+                Hint(
+                    "How many expert reads are in flight at once. More helps only until the storage " +
+                        "itself is saturated, and past that point it costs without buying anything."
                 )
                 SwitchRow(
-                    "Direct I/O (O_DIRECT)",
-                    "Read experts straight from flash, so the system does not keep a second copy of what the cache above already holds. Falls back automatically where unsupported",
+                    "Read straight from storage",
+                    "Skip the system's file cache, so it does not keep a second copy of what the " +
+                        "expert cache already holds. Falls back on its own where the storage does " +
+                        "not support it.",
                     current.oDirect, enabled = stream,
                 ) { onChange(current.copy(oDirect = it)) }
                 SwitchRow(
-                    "I/O–compute overlap",
-                    "Start the next reads while the current layer is still computing, so waiting for flash happens behind the work instead of after it",
+                    "Read while computing",
+                    "Start the next reads while the current layer is still being computed, so " +
+                        "waiting for storage happens behind the work instead of after it.",
                     current.overlap, enabled = stream,
                 ) { onChange(current.copy(overlap = it)) }
                 LabeledDropdown(
-                    "Dense weights",
+                    "Always-needed weights",
                     DenseWeights.values().map { it.label },
                     current.denseWeights.ordinal,
                     enabled = stream,
                 ) { onChange(current.copy(denseWeights = DenseWeights.values()[it])) }
-                Text(
-                    current.denseWeights.blurb,
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                IntSetting(
-                    "Temporal prefetch (layers)", AppSettings.PREFETCH_CHOICES, current.prefetchLayers,
-                    format = { if (it == 0) "off" else "$it" },
-                    // Mutually exclusive with predictive prefetch: two predictors would speculate
-                    // the same future twice, and the engine refuses the pair.
-                    enabled = stream && cacheOn && !current.predictPrefetch && current.routeAhead == 0,
-                ) { onChange(current.copy(prefetchLayers = it)) }
-                Text(
-                    "Experimental. Bets a layer will reuse the experts it picked for the previous token, and " +
-                        "reads them ahead on lanes that would otherwise sit idle. Needs the cache.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                SwitchRow(
-                    "Predictive prefetch (experimental)",
-                    "Instead of betting on the previous token, asks the next layer's own router one " +
-                        "layer early to find out which experts it will actually want. Reads ahead within " +
-                        "the budget below and keeps what the prediction names. Needs the cache; replaces " +
-                        "temporal prefetch.",
-                    current.predictPrefetch,
-                    enabled = stream && cacheOn && current.prefetchLayers == 0 && current.routeAhead == 0,
-                ) { onChange(current.copy(predictPrefetch = it)) }
-                if (current.predictPrefetch) {
+                Hint(current.denseWeights.blurb)
+
+                ExperimentalGroup {
                     IntSetting(
-                        "Predicted misses to read ahead", AppSettings.PREDICT_SPEC_CHOICES, current.predictSpecMax,
-                        format = {
-                            when (it) {
-                                0 -> "0 — retention only (default)"
-                                else -> "$it"
-                            }
-                        },
-                        enabled = stream && cacheOn,
-                    ) { onChange(current.copy(predictSpecMax = it)) }
-                    Text(
-                        "How much reading ahead the prediction may pay for. At zero it reads nothing and only " +
-                            "keeps the experts it names from being evicted — the safe setting, since reading " +
-                            "ahead competes for the same flash the current token is waiting on.",
-                        fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
+                        "Read ahead from the last token", AppSettings.PREFETCH_CHOICES, current.prefetchLayers,
+                        format = { if (it == 0) "Off" else "$it layers" },
+                        // Mutually exclusive with predictive prefetch: two predictors would speculate
+                        // the same future twice, and the engine refuses the pair.
+                        enabled = stream && cacheOn && !current.predictPrefetch && current.routeAhead == 0,
+                    ) { onChange(current.copy(prefetchLayers = it)) }
+                    Hint(
+                        "Bets that a layer will want the same experts it wanted for the previous " +
+                            "token, and fetches them on reads that would otherwise sit idle. Needs " +
+                            "the expert cache."
                     )
+                    SwitchRow(
+                        "Ask the next layer what it wants",
+                        "Instead of betting on the previous token, asks the next layer itself which " +
+                            "experts it is about to need, and fetches those. Needs the expert cache, " +
+                            "and replaces the guess above.",
+                        current.predictPrefetch,
+                        enabled = stream && cacheOn && current.prefetchLayers == 0 && current.routeAhead == 0,
+                    ) { onChange(current.copy(predictPrefetch = it)) }
+                    if (current.predictPrefetch) {
+                        IntSetting(
+                            "How much to fetch on that answer", AppSettings.PREDICT_SPEC_CHOICES,
+                            current.predictSpecMax,
+                            format = { if (it == 0) "Only keep what it names" else "$it per layer" },
+                            enabled = stream && cacheOn,
+                        ) { onChange(current.copy(predictSpecMax = it)) }
+                        Hint(
+                            "At the first setting it fetches nothing and only protects the experts it " +
+                                "names from being thrown out, which is the safe choice: fetching ahead " +
+                                "competes for the same storage the current token is waiting on."
+                        )
+                    }
                 }
             }
 
-            Section("Speed / quality") {
-                // Speculation sits at the top of this section because it is the only entry here that
-                // does NOT trade quality: it changes how many tokens a decode confirms, not what the
-                // model computes. Not gated on the streamer — it is a decode-loop change, so it
-                // applies to the mmap baseline too.
-                LabeledDropdown(
-                    "Guess ahead",
-                    listOf("Off", "Model's own head (MTP)", "Repeated text (n-gram)"),
-                    AppSettings.SPEC_CHOICES.indexOf(current.spec).coerceAtLeast(0),
-                ) { onChange(current.copy(spec = AppSettings.SPEC_CHOICES[it])) }
-                Text(
-                    "Guess the next few tokens, then check the whole group in one pass and keep only what " +
-                        "the model itself would have produced. Nothing is skipped or approximated. The gain " +
-                        "is reading the weights once for several tokens instead of once each; the cost is " +
-                        "that checking several at a time makes every layer touch more experts.\n\n" +
-                        "The head is a small extra part of the model trained to guess — accurate, but only " +
-                        "some models carry it, and running it costs a pass of its own. The n-gram guess " +
-                        "instead looks for the text repeating itself, which costs nothing at all and works " +
-                        "on every model, but only has something to say when the model is quoting or " +
-                        "editing. When it has nothing, that token runs exactly as if this were off.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                if (current.spec != AppSettings.SPEC_OFF) {
-                    IntSetting(
-                        "Tokens guessed per pass", AppSettings.MTP_DRAFT_CHOICES, current.mtpDraft,
-                    ) { onChange(current.copy(mtpDraft = it)) }
-                    Text(
-                        "How far ahead to guess. Further means more tokens confirmed per pass, but the " +
-                            "guesses grow less reliable and a wrong one is paid for and thrown away. " +
-                            "The best setting is rarely the largest.",
-                        fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-                if (current.spec == AppSettings.SPEC_MTP) {
-                    IntSetting(
-                        "Guess only when confident", AppSettings.MTP_P_MIN_CHOICES, current.mtpPMinPct,
-                        format = { if (it == 0) "Always guess" else "Above $it%" },
-                    ) { onChange(current.copy(mtpPMinPct = it)) }
-                    Text(
-                        "Stop guessing as soon as the model is unsure, instead of always filling the pass. " +
-                            "A guess not made costs nothing and keeps the pass narrow, so fewer experts have " +
-                            "to be read — but it also gives up the tokens that guess might have won.",
-                        fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-                // Active-expert (top-k) override is a load-time kv_override, valid in both streaming
-                // and mmap mode, so it is not gated on the streamer.
+            Section("Speed and quality") {
+                // Both entries here are load-time or residency policies, not decode-loop changes.
                 IntSetting(
-                    "Active experts (top-k)", AppSettings.N_EXPERT_CHOICES, current.nExpertUsed,
+                    "Skip experts that are not in memory", AppSettings.DROP_COLD_CHOICES, current.dropColdPct,
                     format = {
                         when (it) {
-                            0 -> "Model default"
-                            else -> "$it"
-                        }
-                    },
-                ) { onChange(current.copy(nExpertUsed = it)) }
-                Text(
-                    "Consult fewer experts per token than the model asks for. Cuts both the computing and " +
-                        "the reading, and changes the reply — a deliberate trade of quality for speed.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                IntSetting(
-                    "Route-ahead (layers)", AppSettings.ROUTE_AHEAD_CHOICES, current.routeAhead,
-                    format = { if (it == 0) "off" else "$it" },
-                    // Excludes both prefetchers (the engine refuses speculating a future this has
-                    // already fixed) and guessing ahead, where a verify pass is several tokens wide
-                    // and this declines to commit on all of them while still paying for itself.
-                    // Needs streaming; the cache is what turns it into early reads.
-                    enabled = !current.mmap && current.prefetchLayers == 0 && !current.predictPrefetch &&
-                        current.spec == AppSettings.SPEC_OFF,
-                ) { onChange(current.copy(routeAhead = it)) }
-                Text(
-                    "Experimental, changes the output. Each layer's expert choice is committed N layers " +
-                        "early — so with the cache on their reads start that early and are never wasted. " +
-                        "~20% of choices differ from the router's at 1 layer; quality held in the host A/B.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                IntSetting(
-                    "Drop cold experts (% of even share)", AppSettings.DROP_COLD_CHOICES, current.dropColdPct,
-                    // The rung labels carry the trade, so the blurb below does not have to repeat
-                    // it: which value to pick is the only real question this setting poses.
-                    format = {
-                        when (it) {
-                            0 -> "off"
-                            50 -> "50% — barely bites"
-                            75 -> "75% — recommended"
-                            100 -> "100% — fastest, roughest"
+                            0 -> "Off"
+                            50 -> "Rarely"
+                            75 -> "Recommended"
+                            100 -> "Fastest, roughest"
                             else -> "$it%"
                         }
                     },
                     // Unlike top-k, this one asks the expert source what is resident, so it needs
-                    // both the streamer and a live cache — the same condition prefetch is under.
-                    enabled = !current.mmap &&
-                        (current.cacheMb == AppSettings.CACHE_AUTO || current.cacheMb > 0),
+                    // both the streamer and a live cache, the same condition prefetch is under.
+                    enabled = stream && cacheOn,
                 ) { onChange(current.copy(dropColdPct = it)) }
-                Text(
-                    "Waiting for an expert that is not already in memory is what slows a token down. This " +
-                        "skips one — but only when it is missing AND the router barely wanted it, below the " +
-                        "chosen share of an even split. Experts already in memory always run, and the " +
-                        "strongest is never skipped, so quality is spent only where it buys back a read." +
-                        "\n\nThe reply changes, and unlike Active experts not the same way twice: what gets " +
-                        "skipped depends on what the cache happened to be holding.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
+                Hint(
+                    "Waiting for an expert that is not already in memory is what slows a token down. " +
+                        "This skips one, but only when it is missing and the model barely wanted it. " +
+                        "An expert already in memory always runs, and the one the model wanted most " +
+                        "is never skipped, so quality is given up only where it buys back a read." +
+                        "\n\nThe reply changes, and not the same way twice: what gets skipped depends " +
+                        "on what the cache happened to be holding at that moment."
                 )
-                // The threshold is a share of 1/top-k, so a narrow routing changes what the same
-                // percentage means: 75% is "below 9.4%" at eight experts but "below 18.8%" at four.
-                // Only shown once a model is loaded and reports its width — guessing would be worse
-                // than saying nothing.
+                // The threshold is a share of the even split, so a narrow routing changes what the
+                // same setting means. Only shown once a model is loaded and reports its width;
+                // guessing would be worse than saying nothing.
                 val topk = ui.nExpertUsed
                 if (current.dropColdPct > 0 && topk != null && topk in 1..4) {
                     Text(
-                        "⚠ This model routes only $topk experts per token, so the same share covers far more " +
-                            "of the reply than it does on a model that routes many. Check the answers, or " +
-                            "turn this off for this model.",
+                        "This model asks for very few experts per token, so the same setting covers " +
+                            "far more of the reply than it does on a model that asks for many. Check " +
+                            "the answers, or turn this off for this model.",
                         fontSize = 12.sp, color = MaterialTheme.colorScheme.error,
+                    )
+                }
+                IntSetting(
+                    "Experts per token", AppSettings.N_EXPERT_CHOICES, current.nExpertUsed,
+                    format = { if (it == 0) "As the model asks" else "$it" },
+                ) { onChange(current.copy(nExpertUsed = it)) }
+                Hint(
+                    "Consult fewer experts per token than the model asks for. Cuts the computing and " +
+                        "the reading together, and changes the reply: a deliberate trade of quality " +
+                        "for speed."
+                )
+
+                ExperimentalGroup {
+                    LabeledDropdown(
+                        "Guess several tokens at once",
+                        listOf("Off", "The model's own guess", "Text that repeats"),
+                        AppSettings.SPEC_CHOICES.indexOf(current.spec).coerceAtLeast(0),
+                        // Excluded by route-ahead, which declines to commit across a wider pass while
+                        // still paying for its prediction.
+                        enabled = current.routeAhead == 0,
+                    ) { onChange(current.copy(spec = AppSettings.SPEC_CHOICES[it])) }
+                    Hint(
+                        "Guess the next few tokens, then check the whole group in one pass and keep " +
+                            "only what the model itself would have produced. Nothing is skipped or " +
+                            "approximated, so the reply is the one you would have got anyway. It wins " +
+                            "by reading the weights once for several tokens instead of once each, and " +
+                            "loses when checking several at a time makes every layer touch more " +
+                            "experts than it would have." +
+                            "\n\nThe model's own guess comes from a small extra part trained for it: " +
+                            "accurate, but only some models carry it and running it costs a pass of " +
+                            "its own. The other watches for the text repeating itself, costs nothing " +
+                            "and works on every model, but has something to say only when the model " +
+                            "is quoting or editing. When it has nothing, that token runs exactly as " +
+                            "if this were off."
+                    )
+                    if (current.spec != AppSettings.SPEC_OFF) {
+                        IntSetting(
+                            "Tokens per guess", AppSettings.MTP_DRAFT_CHOICES, current.mtpDraft,
+                        ) { onChange(current.copy(mtpDraft = it)) }
+                        Hint(
+                            "How far ahead to guess. Further means more tokens confirmed per pass, but " +
+                                "guesses grow less reliable the further out they go, and a wrong one is " +
+                                "paid for and thrown away. The best setting is rarely the largest."
+                        )
+                    }
+                    if (current.spec == AppSettings.SPEC_MTP) {
+                        IntSetting(
+                            "Guess only when sure", AppSettings.MTP_P_MIN_CHOICES, current.mtpPMinPct,
+                            format = { if (it == 0) "Always guess" else "Above $it%" },
+                        ) { onChange(current.copy(mtpPMinPct = it)) }
+                        Hint(
+                            "Stop guessing as soon as the model is unsure, instead of always filling " +
+                                "the pass. A guess not made costs nothing and keeps the pass narrow, so " +
+                                "fewer experts have to be read, but it also gives up the tokens that " +
+                                "guess might have won."
+                        )
+                    }
+                    IntSetting(
+                        "Decide the experts early", AppSettings.ROUTE_AHEAD_CHOICES, current.routeAhead,
+                        format = { if (it == 0) "Off" else "$it layers early" },
+                        // Excludes both prefetchers (the engine refuses speculating a future this has
+                        // already fixed) and guessing ahead. Needs streaming; the cache is what turns
+                        // it into early reads.
+                        enabled = stream && current.prefetchLayers == 0 && !current.predictPrefetch &&
+                            current.spec == AppSettings.SPEC_OFF,
+                    ) { onChange(current.copy(routeAhead = it)) }
+                    Hint(
+                        "Settles which experts a layer will use before that layer is reached, so their " +
+                            "reads can start early and none of them is ever wasted. Changes the reply: " +
+                            "some of those choices differ from the ones the model would have made. " +
+                            "Cannot be combined with guessing ahead or with the fetching above."
                     )
                 }
             }
@@ -282,16 +280,15 @@ fun SettingsScreen(current: AppSettings, onChange: (AppSettings) -> Unit, onBack
                 IntSetting("Tokens to generate", AppSettings.NPREDICT_CHOICES, current.nPredict) {
                     onChange(current.copy(nPredict = it))
                 }
-                IntSetting("Context (tokens)", AppSettings.CTX_CHOICES, current.sessionCtx) {
+                IntSetting("Conversation length", AppSettings.CTX_CHOICES, current.sessionCtx) {
                     onChange(current.copy(sessionCtx = it))
                 }
-                Text(
-                    "How much of the conversation the session can hold, prompt and reply together. " +
-                        "It is also memory: the KV cache is sized for it once when the model opens, " +
-                        "and on a model that already fills RAM that memory comes out of the expert " +
-                        "cache and the weights. Shorten it on the largest models, raise it when you " +
-                        "want long conversations and have the room. Changing it reopens the session.",
-                    fontSize = 12.sp, color = MaterialTheme.colorScheme.onSurfaceVariant,
+                Hint(
+                    "How much of the conversation the model can hold at once, your prompt and its " +
+                        "reply together. It is also memory: room for it is set aside when the model " +
+                        "opens, and on a model that already fills your device that room comes out of " +
+                        "the expert cache. Shorten it on the largest models, raise it when you want " +
+                        "long conversations and have the space. Changing it reopens the model."
                 )
             }
 
@@ -299,13 +296,13 @@ fun SettingsScreen(current: AppSettings, onChange: (AppSettings) -> Unit, onBack
                 SwitchRow(
                     "Thinking",
                     if (thinkingLocked)
-                        "This model always reasons — it offers no way to turn thinking off, so the " +
-                            "switch is disabled here instead of being ignored silently. Its reasoning " +
-                            "still shows in a collapsible block above the reply."
+                        "This model always reasons and offers no way to turn it off, so the switch " +
+                            "is disabled here rather than ignored silently. Its reasoning still " +
+                            "appears in a block above the reply."
                     else
-                        "Let a reasoning model think before answering; its reasoning shows in a " +
-                            "collapsible block above the reply. Off tells the model to skip thinking. " +
-                            "No effect on models that don't reason.",
+                        "Let a model that can reason think before answering; its reasoning appears " +
+                            "in a block above the reply. Off asks it to answer directly. No effect " +
+                            "on models that do not reason.",
                     // Locked reads ON, not OFF: the model reasons on every turn, and that is what
                     // the switch should be showing whatever the stored preference says.
                     checked = current.thinking || thinkingLocked,
@@ -315,9 +312,10 @@ fun SettingsScreen(current: AppSettings, onChange: (AppSettings) -> Unit, onBack
 
             Section("Diagnostics") {
                 SwitchRow(
-                    "Metrics CSV",
-                    "Write one CSV per session: every token's timings, its page faults, the cache budget " +
-                        "and where memory sat. Takes effect on the next session; share it from the menu",
+                    "Save a metrics file",
+                    "Write one file per session with every token's timings, the memory it needed and " +
+                        "where the time went. Takes effect the next time a model opens; share it from " +
+                        "the menu.",
                     current.metricsCsv,
                 ) { onChange(current.copy(metricsCsv = it)) }
             }
